Remove `isIdentifier` function.
The logic here is inverted, previously we checked if input is an identifier, and
if it wasn't, we coerced it into an url. Now that we support search
functionality, we check if input looks like an url, and if it doesn't we assume
it's either identifier or a search string, and since these go to the same
endpoint, there is no need to check if it actually is an identifier.