Add "Suppress Author" checkbox to copy citation dialog (#177)
And fix a bug with previous locator showing up when returning to the
dialog

(We might want to use something less technical than "Suppress Author",
though a default-checked "Include Author" is a little strange. We also
should disable the Copy button when suppressing the author causes there
to be no output (or, if we want to be clever about it, render a
suppressed-author version ahead of time and hide the checkbox if it
would be empty).