Optimise getCitation method #211
Citeproc citations registry is built using `appendCitationCluster` which in turn
calls `processCitationCluster` which makes it computationally expensive. This
needs to be rebuilt every time bibliography is in any way changed.

To optimise this process we only build the registry upon opening the copy
citation dialog, then assume registry is fine because no changes to the
bibliography could have happened while the modal was open.